requirements.txt
fastapi
uvicorn[standard]
httpx
orjson
pydantic
prometheus-client
python-dotenv
tiktoken
python-toon

